+++ src/Controller/itens.controller.js
@@ -1,20 +1,41 @@
 import { Itens } from "../Model/itens.model.js";
 import { insert, updateById, deleteById, getById, getAll,getSomeById } from "../crud.js";
+import { validationResult } from 'express-validator';
 
 class ItensController {
   static async getAllItens(req, res) {
     res.json(await getAll(Itens));
   }
   static async getItemById(req, res) {
-    res.json(await getById(req.body, Itens));
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    res.json(await getById(req.params, Itens));
   }
   static async postItem(req, res) {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
     res.json(await insert(req.body, Itens));
   }
   static async putItem(req, res) {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
     res.json(await updateById(req.body, Itens));
   }
   static async deleteItem(req, res) {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
     res.json(await deleteById(req.body, Itens));
   }
 }
